+++ omnibenchmark/cli/storage.py
@@ -16,6 +16,9 @@ from omnibenchmark.io.files import checksum_files
 from omnibenchmark.io.files import list_files
 from omnibenchmark.io.files import download_files
 from omnibenchmark.io.tree import tree_string_from_list
+from packaging.version import Version
+from datetime import datetime
+from difflib import unified_diff
 
 if TYPE_CHECKING:
     from omnibenchmark.io.MinIOStorage import MinIOStorage
@@ -68,8 +71,38 @@ def storage(ctx):
     ctx.ensure_object(dict)
 
 
+@click.group(name="files")
+@click.pass_context
+def files(ctx):
+    """Manage files in remote storage."""
+    ctx.ensure_object(dict)
+
+
+storage.add_command(files)
+
+
+@click.group(name="version")
+@click.pass_context
+def version(ctx):
+    """Manage benchmark versions."""
+    ctx.ensure_object(dict)
+
+
+storage.add_command(version)
+
+
+@click.group(name="policy")
+@click.pass_context
+def policy(ctx):
+    """Manage storage policies."""
+    ctx.ensure_object(dict)
+
+
+storage.add_command(policy)
+
+
 @add_debug_option
-@storage.command(name="create-version")
+@version.command(name="create")
 @click.option(
     "-b",
     "--benchmark",
@@ -97,7 +130,7 @@ def create_benchmark_version(benchmark: str):
 
 
 @add_debug_option
-@storage.command(name="list")
+@files.command(name="list")
 @click.option(
     "-b",
     "--benchmark",
@@ -139,7 +172,7 @@ def list_all_files(
 
 
 @add_debug_option
-@storage.command(name="download")
+@files.command(name="download")
 @click.option(
     "-b",
     "--benchmark",
@@ -200,7 +233,7 @@ def download_all_files(
 
 
 @add_debug_option
-@storage.command(name="checksum")
+@files.command(name="checksum")
 @click.option(
     "-b",
     "--benchmark",
@@ -233,7 +266,8 @@ def checksum_all_files(benchmark: str):
     logger.info("Done")
 
 
-@storage.command(name="create-policy")
+@add_debug_option
+@policy.command(name="create")
 @click.option(
     "-b",
     "--benchmark",
@@ -261,6 +295,7 @@ def create_policy(benchmark_path: str):
         raise click.Abort()
 
 
+@add_debug_option
 @storage.command("archive")
 @click.option(
     "--benchmark",
@@ -339,14 +374,14 @@ def archive_benchmark(
     local_storage,
     out_dir,
 ):
+    """Archive a benchmark and its artifacts."""
+
     # Validate out_dir usage
     if not local_storage and out_dir != "out":
         logger.error(
             "-Invalid arguments: --out-dir can only be used with --local_storage"
         )
         sys.exit(1)
-
-    """Archive a benchmark"""
 
     assert benchmark is not None
 
@@ -381,3 +416,133 @@ def archive_benchmark(
         click.echo(f"Files to archive:\n{tree_string_from_list(archive_file)}")
     else:
         click.echo(f"Created archive: {archive_file}")
+
+
+@add_debug_option
+@version.command("diff")
+@click.option(
+    "--benchmark",
+    "-b",
+    "benchmark_path",
+    required=True,
+    type=click.Path(exists=True),
+    help="Path to benchmark yaml file or benchmark id.",
+    envvar="OB_BENCHMARK",
+)
+@click.option(
+    "--version1",
+    "-v1",
+    required=True,
+    type=str,
+    help="Reference version.",
+)
+@click.option(
+    "--version2",
+    "-v2",
+    required=True,
+    type=str,
+    help="Version to compare with.",
+)
+@click.pass_context
+def diff_benchmark(ctx, benchmark_path: str, version1, version2):
+    """Show differences between 2 benchmark versions."""
+    from omnibenchmark.io.storage import get_storage, remote_storage_args
+
+    logger.info(
+        f"Found the following differences in {benchmark_path} for {version1} and {version2}."
+    )
+    b = BenchmarkExecution(Path(benchmark_path))
+    auth_options = remote_storage_args(benchmark_path)
+
+    api = b.get_storage_api()
+    bucket = b.get_storage_bucket_name()
+    if api is None or bucket is None:
+        raise (ValueError)
+    # setup storage
+    #
+    # TODO: use walrus
+    ss = get_storage(api, auth_options, bucket)
+    if ss is not None:
+        # get objects for first version
+        ss.set_version(version1)
+        ss._get_objects()
+        files_v1 = [
+            f"{f[0]}   {f[1]['size']}   {datetime.fromisoformat(f[1]['last_modified']).strftime('%Y-%m-%d %H:%M:%S')}\n"
+            for f in ss.files.items()
+        ]
+        if f"versions/{version1}.csv" in ss.files.keys():
+            creation_time_v1 = datetime.fromisoformat(
+                ss.files[f"versions/{version1}.csv"]["last_modified"]
+            ).strftime("%Y-%m-%d %H:%M:%S")
+        else:
+            creation_time_v1 = ""
+
+        # get objects for second version
+        ss.set_version(version2)
+        ss._get_objects()
+        files_v2 = [
+            f"{f[0]}   {f[1]['size']}   {datetime.fromisoformat(f[1]['last_modified']).strftime('%Y-%m-%d %H:%M:%S')}\n"
+            for f in ss.files.items()
+        ]
+        if f"versions/{version2}.csv" in ss.files.keys():
+            creation_time_v2 = datetime.fromisoformat(
+                ss.files[f"versions/{version2}.csv"]["last_modified"]
+            ).strftime("%Y-%m-%d %H:%M:%S")
+        else:
+            creation_time_v2 = ""
+
+    # diff the two versions
+    click.echo(
+        "".join(
+            list(
+                unified_diff(
+                    files_v1,
+                    files_v2,
+                    fromfile=f"version {version1}",
+                    tofile=f"version {version2}",
+                    fromfiledate=creation_time_v1,
+                    tofiledate=creation_time_v2,
+                )
+            )
+        )
+    )
+
+
+@add_debug_option
+@version.command("list")
+@click.option(
+    "--benchmark",
+    "-b",
+    "benchmark_path",
+    required=True,
+    type=click.Path(exists=True),
+    help="Path to benchmark yaml file or benchmark id.",
+    envvar="OB_BENCHMARK",
+)
+@click.pass_context
+def list_versions(ctx, benchmark_path: str):
+    """List all available benchmark versions."""
+    from omnibenchmark.io.storage import get_storage, remote_storage_args
+
+    logger.info(f"Available versions of {benchmark_path}:")
+
+    b = BenchmarkExecution(Path(benchmark_path))
+    auth_options = remote_storage_args(b)
+    api = b.get_storage_api()
+    bucket = b.get_storage_bucket_name()
+
+    if api is None:
+        raise ValueError("No storage API found")
+    if bucket is None:
+        raise ValueError("No storage bucket found")
+
+    # setup storage
+    ss = get_storage(api, auth_options, bucket)
+    if ss is None:
+        raise ValueError("No storage found")
+
+    if len(ss.versions) > 0:
+        if len(ss.versions) > 1:
+            ss.versions.sort(key=Version)
+        for version in ss.versions:
+            click.echo(f"{str(version):>8}")